Custom check_admin function + other cleanup
